- delete per-route error.tsx (R1-C1) — caught by existing app/(main)/error.tsx boundary - move server actions from prisma/services/ to prisma/actions/ (R1-M2) - merge query functions into prisma/data/positions.ts alongside getOpenPositions (R1-M2, R1-N1) - fix all action return shapes from {ok} to void/{error}/data (R1-M1) and update all call-sites - move PositionManager, PositionForEdit, PositionQuestionForEdit to lib/types.ts (R1-M4) - add shadcn Select component and replace raw <select> elements in all three forms (R1-M3) - move STATUS_*/QUESTION_TYPE_* constants to lib/constants.ts (R1-L1) - align create dialog date inputs to datetime-local (R1-L2) - run prettier:fix across repo (pre-existing formatting debt)
