+++ lib/types.ts
@@ -24,6 +24,38 @@ export type PositionWithQuestions = {
   }[];
 };
 
+export type PositionManager = {
+  id: string;
+  name: string | null;
+  email: string;
+};
+
+// Narrowed question shape — only the fields rendered by the edit page and
+// PositionQuestionsSection; audit columns are excluded to avoid leaking them
+// across the server/client prop boundary.
+export type PositionQuestionForEdit = {
+  id: string;
+  positionId: string;
+  label: string;
+  type: QuestionType;
+  required: boolean;
+  options: string[];
+  order: number;
+};
+
+// Only the six position fields consumed on the edit page; audit columns are
+// excluded so they are never serialized across the server/client boundary.
+export type PositionForEdit = {
+  id: string;
+  title: string;
+  description: string;
+  status: PositionStatus;
+  opensAt: Date | null;
+  closesAt: Date | null;
+  questions: PositionQuestionForEdit[];
+  managers: PositionManager[];
+};
+
 export type DraftApplication = Application & {
   globalAnswers: GlobalApplicationAnswer[];
   positionAnswers: PositionApplicationAnswer[];
