+++ package.json
@@ -32,6 +32,7 @@
     "@radix-ui/react-label": "^2.1.10",
     "@radix-ui/react-popover": "^1.1.17",
     "@radix-ui/react-radio-group": "^1.4.1",
+    "@radix-ui/react-select": "^2.3.1",
     "@radix-ui/react-slot": "^1.3.0",
     "@radix-ui/react-tabs": "^1.1.15",
     "@radix-ui/react-tooltip": "^1.2.10",
