+++ prisma/actions/position-question-actions.ts
@@ -53,19 +53,16 @@ async function checkAccess(
 
 export async function createPositionQuestion(
   input: unknown,
-): Promise<
-  | { ok: true; data: { id: string; order: number } }
-  | { ok: false; error: string }
-> {
+): Promise<{ id: string; order: number } | { error: string }> {
   const user = await getCurrentUser();
 
   const parsed = createPositionQuestionSchema.safeParse(input);
-  if (!parsed.success) return { ok: false, error: 'Invalid input' };
+  if (!parsed.success) return { error: 'Invalid input' };
 
   const { positionId, label, type, required, options } = parsed.data;
 
   const hasAccess = await checkAccess(positionId, user.id, user.isAdmin);
-  if (!hasAccess) return { ok: false, error: 'Unauthorized' };
+  if (!hasAccess) return { error: 'Unauthorized' };
 
   const maxOrder = await prisma.positionQuestion.aggregate({
     where: { positionId, deletedAt: null },
@@ -88,21 +85,21 @@ export async function createPositionQuestion(
   });
 
   revalidatePath(`/positions/${positionId}/edit`);
-  return { ok: true, data: { id: created.id, order: created.order } };
+  return { id: created.id, order: created.order };
 }
 
 export async function updatePositionQuestion(
   input: unknown,
-): Promise<{ ok: true } | { ok: false; error: string }> {
+): Promise<void | { error: string }> {
   const user = await getCurrentUser();
 
   const parsed = updatePositionQuestionSchema.safeParse(input);
-  if (!parsed.success) return { ok: false, error: 'Invalid input' };
+  if (!parsed.success) return { error: 'Invalid input' };
 
   const { id, positionId, label, type, required, options } = parsed.data;
 
   const hasAccess = await checkAccess(positionId, user.id, user.isAdmin);
-  if (!hasAccess) return { ok: false, error: 'Unauthorized' };
+  if (!hasAccess) return { error: 'Unauthorized' };
 
   // Scope the write to positionId to prevent IDOR across positions
   const result = await prisma.positionQuestion.updateMany({
@@ -110,24 +107,23 @@ export async function updatePositionQuestion(
     data: { label, type, required, options, updatedById: user.id },
   });
 
-  if (result.count === 0) return { ok: false, error: 'Not found' };
+  if (result.count === 0) return { error: 'Not found' };
 
   revalidatePath(`/positions/${positionId}/edit`);
-  return { ok: true };
 }
 
 export async function deletePositionQuestion(
   input: unknown,
-): Promise<{ ok: true } | { ok: false; error: string }> {
+): Promise<void | { error: string }> {
   const user = await getCurrentUser();
 
   const parsed = deletePositionQuestionSchema.safeParse(input);
-  if (!parsed.success) return { ok: false, error: 'Invalid input' };
+  if (!parsed.success) return { error: 'Invalid input' };
 
   const { id, positionId } = parsed.data;
 
   const hasAccess = await checkAccess(positionId, user.id, user.isAdmin);
-  if (!hasAccess) return { ok: false, error: 'Unauthorized' };
+  if (!hasAccess) return { error: 'Unauthorized' };
 
   // Scope the write to positionId to prevent IDOR across positions
   const result = await prisma.positionQuestion.updateMany({
@@ -135,8 +131,7 @@ export async function deletePositionQuestion(
     data: { deletedAt: new Date(), deletedById: user.id },
   });
 
-  if (result.count === 0) return { ok: false, error: 'Not found' };
+  if (result.count === 0) return { error: 'Not found' };
 
   revalidatePath(`/positions/${positionId}/edit`);
-  return { ok: true };
 }
